requirements.txt
pandas==3.0.1
openpyxl==3.1.5
python-pptx==1.0.2
lxml==6.0.2
